+++ cmd/compute/elastic_ip/elastic_ip_list.go
@@ -50,13 +50,20 @@ func (c *elasticIPListCmd) CmdPreRun(cmd *cobra.Command, args []string) error {
 }
 
 func (c *elasticIPListCmd) CmdRun(_ *cobra.Command, _ []string) error {
-	var zones []string
+	var zones []v3.ZoneName
 	ctx := exocmd.GContext
 
 	if c.Zone != "" {
-		zones = []string{c.Zone}
+		zones = []v3.ZoneName{v3.ZoneName(c.Zone)}
 	} else {
-		zones = utils.AllZones
+		client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(c.Zone))
+		if err != nil {
+			return err
+		}
+		zones, err = utils.AllZonesV3(ctx, *client)
+		if err != nil {
+			return err
+		}
 	}
 
 	out := make(elasticIPListOutput, 0)
@@ -69,7 +76,7 @@ func (c *elasticIPListCmd) CmdRun(_ *cobra.Command, _ []string) error {
 		}
 		done <- struct{}{}
 	}()
-	err := utils.ForEachZone(zones, func(zone string) error {
+	err := utils.ForEachZone(zones, func(zone v3.ZoneName) error {
 		client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(zone))
 		if err != nil {
 			return err
@@ -85,7 +92,7 @@ func (c *elasticIPListCmd) CmdRun(_ *cobra.Command, _ []string) error {
 				res <- elasticIPListItemOutput{
 					ID:        e.ID.String(),
 					IPAddress: e.IP,
-					Zone:      zone,
+					Zone:      string(zone),
 				}
 			}