+++ cmd/compute/instance/instance_private_network_detach.go
@@ -1,18 +1,15 @@
 package instance
 
 import (
-	"errors"
 	"fmt"
 	"strings"
 
 	"github.com/spf13/cobra"
 
 	exocmd "github.com/exoscale/cli/cmd"
-	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
 	"github.com/exoscale/cli/pkg/output"
 	"github.com/exoscale/cli/utils"
-	exoapi "github.com/exoscale/egoscale/v2/api"
 	v3 "github.com/exoscale/egoscale/v3"
 )
 
@@ -47,19 +44,37 @@ func (c *instancePrivnetDetachCmd) CmdPreRun(cmd *cobra.Command, args []string) 
 }
 
 func (c *instancePrivnetDetachCmd) CmdRun(_ *cobra.Command, _ []string) error {
-	ctx := exoapi.WithEndpoint(exocmd.GContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, c.Zone))
-
-	instance, err := globalstate.EgoscaleClient.FindInstance(ctx, c.Zone, c.Instance)
+	ctx := exocmd.GContext
+	client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(c.Zone))
 	if err != nil {
-		if errors.Is(err, exoapi.ErrNotFound) {
-			return fmt.Errorf("resource not found in zone %q", c.Zone)
-		}
 		return err
 	}
 
-	privateNetwork, err := globalstate.EgoscaleClient.FindPrivateNetwork(ctx, c.Zone, c.PrivateNetwork)
+	instances, err := client.ListInstances(ctx)
 	if err != nil {
-		return fmt.Errorf("error retrieving Private Network: %w", err)
+		return err
+	}
+	instance, err := instances.FindListInstancesResponseInstances(c.Instance)
+	if err != nil {
+		return err
+	}
+
+	privateNetworks, err := client.ListPrivateNetworks(ctx)
+	if err != nil {
+		return err
+	}
+	privateNetwork, err := privateNetworks.FindPrivateNetwork(c.PrivateNetwork)
+	if err != nil {
+		return err
+	}
+
+	op, err := client.DetachInstanceFromPrivateNetwork(ctx, privateNetwork.ID, v3.DetachInstanceFromPrivateNetworkRequest{
+		Instance: &v3.Instance{
+			ID: instance.ID,
+		},
+	})
+	if err != nil {
+		return err
 	}
 
 	utils.DecorateAsyncOperation(fmt.Sprintf(
@@ -67,9 +82,7 @@ func (c *instancePrivnetDetachCmd) CmdRun(_ *cobra.Command, _ []string) error {
 		c.Instance,
 		c.PrivateNetwork,
 	), func() {
-		if err = globalstate.EgoscaleClient.DetachInstanceFromPrivateNetwork(ctx, c.Zone, instance, privateNetwork); err != nil {
-			return
-		}
+		_, err = client.Wait(ctx, op, v3.OperationStateSuccess)
 	})
 	if err != nil {
 		return err
@@ -78,7 +91,7 @@ func (c *instancePrivnetDetachCmd) CmdRun(_ *cobra.Command, _ []string) error {
 	if !globalstate.Quiet {
 		return (&instanceShowCmd{
 			CliCommandSettings: c.CliCommandSettings,
-			Instance:           *instance.ID,
+			Instance:           instance.ID.String(),
 			Zone:               v3.ZoneName(c.Zone),
 		}).CmdRun(nil, nil)
 	}