+++ cmd/compute/instance/instance_private_network_updateip.go
@@ -1,7 +1,6 @@
 package instance
 
 import (
-	"errors"
 	"fmt"
 	"net"
 	"strings"
@@ -9,11 +8,9 @@ import (
 	"github.com/spf13/cobra"
 
 	exocmd "github.com/exoscale/cli/cmd"
-	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
 	"github.com/exoscale/cli/pkg/output"
 	"github.com/exoscale/cli/utils"
-	exoapi "github.com/exoscale/egoscale/v2/api"
 	v3 "github.com/exoscale/egoscale/v3"
 )
 
@@ -50,31 +47,38 @@ func (c *instancePrivnetUpdateIPCmd) CmdPreRun(cmd *cobra.Command, args []string
 }
 
 func (c *instancePrivnetUpdateIPCmd) CmdRun(_ *cobra.Command, _ []string) error {
-	ctx := exoapi.WithEndpoint(exocmd.GContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, c.Zone))
-
-	instance, err := globalstate.EgoscaleClient.FindInstance(ctx, c.Zone, c.Instance)
+	ctx := exocmd.GContext
+	client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(c.Zone))
 	if err != nil {
-		if errors.Is(err, exoapi.ErrNotFound) {
-			return fmt.Errorf("resource not found in zone %q", c.Zone)
-		}
 		return err
 	}
 
-	privateNetwork, err := globalstate.EgoscaleClient.FindPrivateNetwork(ctx, c.Zone, c.PrivateNetwork)
+	instances, err := client.ListInstances(ctx)
 	if err != nil {
-		return fmt.Errorf("error retrieving Private Network: %w", err)
+		return err
+	}
+	instance, err := instances.FindListInstancesResponseInstances(c.Instance)
+	if err != nil {
+		return err
 	}
 
+	privateNetworks, err := client.ListPrivateNetworks(ctx)
+	if err != nil {
+		return err
+	}
+	privateNetwork, err := privateNetworks.FindPrivateNetwork(c.PrivateNetwork)
+	if err != nil {
+		return err
+	}
+
+	op, err := client.UpdatePrivateNetworkInstanceIP(ctx, privateNetwork.ID, v3.UpdatePrivateNetworkInstanceIPRequest{
+		Instance: &v3.UpdatePrivateNetworkInstanceIPRequestInstance{
+			ID: instance.ID,
+		},
+		IP: net.ParseIP(c.IPAddress),
+	})
 	utils.DecorateAsyncOperation(fmt.Sprintf("Updating instance %q Private Network IP address...", c.Instance), func() {
-		if err = globalstate.EgoscaleClient.UpdatePrivateNetworkInstanceIPAddress(
-			ctx,
-			c.Zone,
-			instance,
-			privateNetwork,
-			net.ParseIP(c.IPAddress),
-		); err != nil {
-			return
-		}
+		_, err = client.Wait(ctx, op, v3.OperationStateSuccess)
 	})
 	if err != nil {
 		return err
@@ -83,7 +87,7 @@ func (c *instancePrivnetUpdateIPCmd) CmdRun(_ *cobra.Command, _ []string) error 
 	if !globalstate.Quiet {
 		return (&instanceShowCmd{
 			CliCommandSettings: c.CliCommandSettings,
-			Instance:           *instance.ID,
+			Instance:           instance.ID.String(),
 			Zone:               v3.ZoneName(c.Zone),
 		}).CmdRun(nil, nil)
 	}