+++ cmd/compute/load_balancer/nlb_list.go
@@ -51,14 +51,20 @@ func (c *nlbListCmd) CmdPreRun(cmd *cobra.Command, args []string) error {
 }
 
 func (c *nlbListCmd) CmdRun(_ *cobra.Command, _ []string) error {
+	var zones []v3.ZoneName
 	ctx := exocmd.GContext
 
-	var zones []string
-
 	if c.Zone != "" {
-		zones = []string{c.Zone}
+		zones = []v3.ZoneName{v3.ZoneName(c.Zone)}
 	} else {
-		zones = utils.AllZones
+		client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(c.Zone))
+		if err != nil {
+			return err
+		}
+		zones, err = utils.AllZonesV3(ctx, *client)
+		if err != nil {
+			return err
+		}
 	}
 
 	out := make(nlbListOutput, 0)
@@ -71,7 +77,7 @@ func (c *nlbListCmd) CmdRun(_ *cobra.Command, _ []string) error {
 		}
 		done <- struct{}{}
 	}()
-	err := utils.ForEachZone(zones, func(zone string) error {
+	err := utils.ForEachZone(zones, func(zone v3.ZoneName) error {
 		client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(zone))
 		if err != nil {
 			return err